requirements.txt
Jinja2==3.0.1
aiofiles==0.7.0
psycopg2-binary==2.9.1
pydantic==1.8.2
python-dotenv==0.19.0
SQLAlchemy==1.4.23
typing-extensions==3.10.0.2
fastapi_pagination
python-multipart